# swagger.json
{
    "openapi": "3.0.0",
    "info":{
        "version":"1.0.0",
        "description": "API for project Postaway. In this we can add posts, comments and like them.",
        "title":"Postaway project API"
    },
    "servers":[
        {
            "url":"http://localhost:3200"
        }
    ],
    "components":{
        "securitySchemes":{
            "JWT":{
                "in":"header",
                "name":"Authorization",
                "type":"apiKey"
            }
        }
    },
    "paths":{
...
        "/api/posts/":{
            "get":{
                "tags":["Posts"],

                "summary":"Get your Posts",
                "description":"This Api helps in extracting all the posts that you posted",
                "security":[
                    {"JWT":{}}
                ],
                "responses":{
                    "200":{
                        "description":"OK"
                    },
                    "401":{
                        "description": "Unauthorized"
                    }
                }
            },

            "post":{
                "tags":["Posts"],
                "summary":"Post",
                "description":"This Api helps in posting",
                "security":[
                    {"JWT":{}}
                ],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "properties":{
                                    "caption":{
                                        "type":"string"
                                    },
                                    "imageUrl":{
                                        "type":"string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"OK"
                    },
                    "400":{
                        "description":"Invalid Credentials"
                    },
                    "404":{
                        "description":"User not found"
                    }
                }
            }
        },
...
    }
}